+++ error.go
@@ -14,7 +14,7 @@ type Error struct {
 	Traces   []Trace  `json:"traces,omitempty"`
 }
 
-func New(code Code, message string, traces []Trace) *Error {
+func New(code Code, message string, traces ...Trace) Error {
 	return &Error{
 		Code:     code,
 		Message:  message,
@@ -46,7 +46,7 @@ func (err Error) Error() string {
 
 func (err Error) Format() string {
 	return fmt.Sprintf(
-		"error code: %d; error description: %s\n\terror: %s; in file: %s; at line: %d",
+		"code id: %d; description: %s\n\terror message: %s; in file: %s; at line: %d",
 		err.Code.ID,
 		err.Code.Description,
 		err.Message,
@@ -61,8 +61,15 @@ func (err Error) FormatWithTraces() string {
 	})
 }
 
-func (err Error) FormatJSON() string {
-	data, _ := json.Marshal(err)
+func (err Error) FormatJSON(indent bool) string {
+	var data []byte
+
+	if indent {
+		data, _ = json.MarshalIndent(err, "", "\t")
+		return string(data)
+	}
+
+	data, _ = json.Marshal(err)
 	return string(data)
 }
 
@@ -79,5 +86,5 @@ func (trace Trace) IntoError() Error {
 }
 
 func (trace Trace) Format() string {
-	return fmt.Sprintf("error: %s; in file: %s; at line: %d", trace.Message, trace.Position.File, trace.Position.Line)
+	return fmt.Sprintf("trace message: %s; in file: %s; at line: %d", trace.Message, trace.Position.File, trace.Position.Line)
 }